refactor: split interview manager into per-session and dashboard modes
Extracted dashboard-specific logic (election, polling, HTTP push/callback)
into dashboard-manager.ts (~480 lines). manager.ts becomes a thin
composition root (~50 lines) that branches on config.

Per-session mode kept inline in composition root — 28 lines of wiring,
not worth its own file.

Refs #647
